requirements.txt
praw
tweepy
openai  # For summarization (or use another LLM)
requests
python-dotenv
